Bug: 95483 - Emails from Webex not displaying properly
If the iframe content (added to the container div) is a table with align set, the height will be improperly calculated.  See W3C Visual formatting model details, http://www.w3.org/TR/CSS21/visudet.html,  
Section 10.6.3 - calculating the container div height will ignore the table, since an align setting is treated as a 'float', i.e. the table is not an in-flow child.

Changing the container div overflow to 'hidden' makes it calculate its height as per sections 10.6.6 and 10.6.7, where all the block children (i.e the table) are included regardless of whether they are in-flow.

http://bugzilla.zimbra.com/show_bug.cgi?id=95483

Copied from Perforce
 Change: 524013