Bug: 94935 - Admin sharing bug Bug: 95661 - Clicking to shared mailbox folder switches focus to mail app
See also Bug 95660 for a more general case.
A completely shared mail account shows up in the calendar tree with a 'shared mailbox' node, that has the other account's calendars under it.  There is currently no way to create a calendar under the shared mailbox node.  This adds a separate action menu for that node (only for the calendar tree, lets not get too generalized here) that allows calendar creation.  If we want to put the effort into it, it can be extended with other operations later.

Also prevents a click on 'shared-mailbox' from transitioning to the mail app.

http://bugzilla.zimbra.com/show_bug.cgi?id=94935

Copied from Perforce
 Change: 524257