Bug 95697
Flagging a message in conversation view doesn't update visual flag indicator


Regression from bug 90268. The solution there overrides ZmItem._parseFlags but the problem is it overrides it with different param type (array of msg nodes instead of flags string) which is not the best pattern and in this case causes this bug since there are other callers of this method.

My solution is to rename ZmConv._parseFlags to ZmConv._parseFlagsFromMsgs, which is now what's called from ZmConv.prototype._loadFromDom.

This works for the case of flagging (and also clicking on the mark read/unread circle, or using the mark unread/read from the toolbar) an entire conversation fine now.

http://bugzilla.zimbra.com/show_bug.cgi?id=95697

Copied from Perforce
 Change: 524765