Bug: 95758 - Offline: Moved messages to another folder revert back.
After a move,  ZmOffline._handleSyncUpdate was ignoring messages that were not in a folder that was configured for offline retention.  So if the message had been in an offline folder and was moved to one that did not support offline, its folder id was not updated.  Since the sync update doesn't have info about a messages's previous folder, have it attempt to delete (from the IndexedDB mail store) the updated messages that are not in an offline folder.  

http://bugzilla.zimbra.com/show_bug.cgi?id=95758

Copied from Perforce
 Change: 525162