Bug 95882
format as HTML/plain-text action icons in compose view are wrong


Don't use icons for the format as HTML and format as plain text actions. Since we were using icons for doc type that are also used in Briefcase, and the briefcase ones were changed, and now they no longer make sense in the context of the compose view, Josh decided it's better to simply remove the icons as they don't help anyway.

http://bugzilla.zimbra.com/show_bug.cgi?id=95882

Copied from Perforce
 Change: 525287