bug: 32283, 55337 - merge accessibility changes back to JUDASPRIEST
At a high level, this work builds on the infrastructure added in
the last merge. With this, the app chooser, the overview and the
main list view should be fully keyboard navigable. The two former
also have the appropriate ARIA tags, and should be accessible as
well.

http://bugzilla.zimbra.com/show_bug.cgi?id=32283
http://bugzilla.zimbra.com/show_bug.cgi?id=55337

Copied from Perforce
 Change: 525564