Bug 95379
Unselecting use prefix from forward does not work


This fixes 2 problem.

1. The problem described in the bug - if the prefix is | and we forward/reply in text mode, unchecking "use prefix", does not remove the prefix.
The solution for this is in ZmComposeView.prototype._preserveQuotedTex, as it was hardcoded there to only remove ">". I changed it to simply remove either ">" or "|".

2. a secondary problem I found:

In nested quotes (text mode) things don't quite work - for example, nested appears like this:

On Friday, September 12, 2003, at 11:23 AM, Jennifer wrote:

> On Friday, September 12, 2003, at 11:23 AM, Steve wrote:
>
>> blah blah blah blah blah
>> blah blah blah

The problem is that then unchecking "use prefix" in the actions menu, does not remove it from the lines that have multiple, at all. You get:

On Friday, September 12, 2003, at 11:23 AM, Jennifer wrote:

On Friday, September 12, 2003, at 11:23 AM, Steve wrote:

>> blah blah blah blah blah
>> blah blah blah

This is wrong.

The reason is a combination of 2 things:
1. We do not add a space after the prefix char if there's another prefix char after it. This is to reserve horizontal space in case of multiple level quotes.
2. when removing the prefix, we look for the prefix char with a space after, and due to (1) there's not space.

The solution is to make the space after the prefix optional in the regex that removes the prefix. That way it would remove the prefix and space (for regular text) and would just remove the prefix for a prefix that has another prefix after it (nested quotes).

http://bugzilla.zimbra.com/show_bug.cgi?id=95379

Copied from Perforce
 Change: 525803