Bug 95688 - Scrolling and scroll bar functionality broken under Contacts tab.
ZmContactSplitView.prototype._sizeChildren has some hardcoded values for adjusting the height. I think it's better to calculate it dynamically (depending on the skin or future changes, hardcoded numbers may not work well).

I encountered an issue with getY(), or actually getLocation() in Dwt. The problem is that the elements with absolute position may not have "top" and "left" assigned, in which case it defaults to "auto" which then turns into NaN after parseInt is called. For now, I'm calling toWindow from _sizeChildren for the problematic element. I suggest we fix getLocation() when we start working on the next release.

Review Board discussion: https://reviewboard.eng.zimbra.com/r/2691/

http://bugzilla.zimbra.com/show_bug.cgi?id=95688

Copied from Perforce
 Change: 525893