Bug 95669 Bug 95946
Moved message don't appear in destination folder for shared mailbox



In bug 85084, I solved a similar problem, with the difference that there it was a shared folder, instead of a folder in a shared mailbox.

My solution there was to use the rid if the folder is remote. However, rid only exists for mountpoints (that's what the server returns), and in this case the mountpoint is the entire mailbox, and not the "sent" folder.

I realize my solution there was an overkill, with added method (getLocalId), while all I had to really do is use the folder's nId property that is set in the ZmOrganizer constructor to be the "local" part (thus "5" for sent in the "sent" folder of that shared mailbox, for example).

So my solution here is to just use nId and remove getLocalId which was not really doing what I thought it was doing for all cases.

P.S. this solves bug 95946 as well.

http://bugzilla.zimbra.com/show_bug.cgi?id=95669

Copied from Perforce
 Change: 525898