Bug 93525
The folder color of an external IMAP account cannot be changed


The solution here is that since we choose to show the special icons for some IMAP folders, based on some simple heuristic (the name of the folder), we should also enable/disable the same actions as for the equivalent system folder we choose.
So now just like you cannot edit properties of Trash, you can't edit properties of "Trash" in the IMAP. Just like you can edit properties of Drafts but can't change colors, same is true for "Drafts" in the IMAP. Etc. This is a larger scope than just related to the color issue.

http://bugzilla.zimbra.com/show_bug.cgi?id=93525

Copied from Perforce
 Change: 525912