Bug 96007
Message from shared mailbox shows tag name as Folder's name i.e 2 for Inbox and 3 for Trash


creating a non local tag (tag that's on a remote item and there's no similarly named one locally), we used the name as both name and ID, but in this case in ZmOrganizer (which is the parent of ZmTag) we override the name of the ID is of a system folder.

Solution is to name is nonLocal_[name] so it does not collide with the system folder IDs.

http://bugzilla.zimbra.com/show_bug.cgi?id=96007

Copied from Perforce
 Change: 526136