Bug: 96037 - Quick add appt: rubber-banding still happens when moving cursor within dialog
Mouse down/up (or down, drag, up) clicks on the ZmCalendarColView are followed by an onClick event on the calendar view.  The mouse up triggers the QuickAdd dialog display.  The following onClick used to not do anything, but now the accessbility changes use it to generate a pair of mouse down & up events.   The mouse  down re-initialized the appt drag, and apparently things were confused enough at that point that the next mouse up didn't clear it. 

I really don't want to touch the DwtControl._mouseEvent changes, so I've simply discarded the onClick listener for the ZmCalColView.  There was only the DwtControl.onClick listener set up, nothing was customized for ZmCalColView, so it seems to be OK to remove.  Tested the QuickAdd and dragging appts.

http://bugzilla.zimbra.com/show_bug.cgi?id=96037

Copied from Perforce
 Change: 526310