Bug: 95516 - IE8 - three dots ('...') overwrapping on the first char on message list
Very strange (well, its IE8).  The css has overflow-text:ellipsis set, but it seems to be applying the ellipsis on the left hand side, overwriting the text.  
I've talked to Josh, and will apply a band-aid for now,  just removing the overflow-text:ellipsis when its IE8; the text gets clipped by a parent div, so it runs rather close to the left icon, but it still seems preferable to the problem stated in the bug.

http://bugzilla.zimbra.com/show_bug.cgi?id=95516

Copied from Perforce
 Change: 526348