Bug 96042
"item nnnn does not exist" error when clicking on an accepted/declined invite message of shared mailbox


The problem is GetAppointmentRequest was setting the apptId to the "local" part only, even if the invite message is for a shared mailbox.

Solution is to set it to accountId:apptId, getting the accountId that is part of the msgId (if it's in the format acctId:msgId).

http://bugzilla.zimbra.com/show_bug.cgi?id=96042

Copied from Perforce
 Change: 526461