Bug 95689
"Move appointment to shared calendar" confirmation confuses to end user when accepting invite from shared mailbox


Removed the message, by undoing that part of bug 43567. The reasoning, which I confirmed with Josh, is that we now default to the correct calendar in the drop-down based on the account the invite/msg belongs to. So if it's the shared one, it's OK and no reason to warn.

Another reason is that we only show in the drop-down calendars that the user has "write" access to (admin/manager), so it should always work.

http://bugzilla.zimbra.com/show_bug.cgi?id=95689

Copied from Perforce
 Change: 526467