Bug 96060
possible exception when composing in new window


Regression from bug 77268, but the real culprit is bug 57561 - there, the code did a couple of wrong things -
(1) assumed if params.msg is passed, then params.msg.nId exists.
(2) defaulted to this._msg.nId - this is wrong since this._msg is leftover from previous compose view and has nothing to do with this. (it also did the same mistake as (1) in regards to assuming if this._msg is set, then this._msg.nId is set).

My fix takes care of both issues. Only use params.msg and make sure params.msg.nId is not empty, and if either are empty, get the next serial ID like before. (Possibly we should default to some constant, but not sure so I don't change that part).

http://bugzilla.zimbra.com/show_bug.cgi?id=96060

Copied from Perforce
 Change: 526470