Bug 96061
Shows error message "No such account exists",while clicking on 'Propose New Time'


Invitation ID (that is used in GetMsgRequest for some reason, for getting "details", see: ZmCalItem.prototype.getDetail, this.message.load), is set to aptId-msgId, e.g. 288-287.

The problem was that for a shared account, it was appending the aptId followed by a dash before the accountId:msgId thing, thus the aptId became part of the accountId. In the specific case in the bug, it became:
265-c6eba6d7-81b4-42ab-b679-d1b702d5b969:264

I guessed and was correct the format in this case should be:

c6eba6d7-81b4-42ab-b679-d1b702d5b969:265-264

This seems to fix everything...

http://bugzilla.zimbra.com/show_bug.cgi?id=96061

Copied from Perforce
 Change: 526696