bug 55337 - don't add raw elements to toolbars
The conversation view features an inline reply view with a toolbar
containing items for "Send", "Cancel" and "Switch to full compose".
The later item was added as a raw A (i.e. link) element, requiring
special code in DwtToolBar for adding non-control children. This
breaks the usual expectations of the DwtComposite API, and complicates
a forthcoming fix.

To increase consistency, we make the "Switch to full compose" link use
a regular button. This changes its appearance from blue text to
being black like the two others. Alternatively, we can retain the
previous functionality by using a DwtLinkButton.

http://bugzilla.zimbra.com/show_bug.cgi?id=55337

Copied from Perforce
 Change: 527121