Bug 95968
Replying in a conversation thread automatically places the scroll bar at the bottom


I'm not sure why initially the textarea scrolls to the bottom. Not sure if it's a new Chrome bug, or if it's something in our code (was not able to find something we do intentionally).

However, this simple fix fixes the "reply" problem, on Chrome. We have the moveCaretToTop code, where we use selection to move the cursor to the top of the textarea. However, moving the cursor does not scroll back up, as it's independent of scrolling. But I can't think of a case we want to move the cursor to the top but not scroll, so what this fix does is also set the textarea element scrollTop to 0, so it also scrolls to top.

http://bugzilla.zimbra.com/show_bug.cgi?id=95968

Copied from Perforce
 Change: 527219