Bug 95929
Additional Accounts shouldn't be marked as folders


Bug 75418 added the icon, as a somewhat hacky solution to the fact there was some extra space due to missing icon.

Anyway this looks OK now. Also, for POP accounts, it's not the same, we just have on folder (either inbox or a dedicated one), and so it has an icon and that's OK. See ZmFolder.prototype.getIcon for where the POPAccount icon is returned for POP datasource folder.

http://bugzilla.zimbra.com/show_bug.cgi?id=95929

Copied from Perforce
 Change: 527234