Bug: 96237: Editor doesn't get change from HTML to TEXT in Compose window
The tinyMCE editor thought that it was hidden (though it showed the toolbar) based on the textArea being hidden.  Thereafter, switching modes made the tinyMCE editor toggle its show/hide, so that it displayed the toolbar when in plain mode and hid it when in HTML mode (i.e. out of sync with our choices).

I've just removed the hide of the textarea.  The display seems to look OK.

Copied from Perforce
 Change: 527797