Bug: 96267 - Shortcut to go to next/previous conv results in JS error
The first part of the fix was easy, since we were trying
to check the page forward/back nav bar buttons that
no longer exist. The second part was tougher. The conv
controller's lack of a current query kept the conv search
from happening, but we had loaded the conv earlier in
setSelectedItem() when faking the double click. That had
a side effect of mixing up the callback chain so that the
conv did not get displayed. The fix is to make sure the
conv controller knows its current query.

http://bugzilla.zimbra.com/show_bug.cgi?id=96267

Copied from Perforce
 Change: 527906