Bug 94476 - Select the contact group via the button "To" is Incomplete .
Changeset:
1. ZmContactPicker.js

Cause ::- The shared members of shared contact group donot come in the list because 'getContactRequest' is not fired, when you chooose the option 'personal and shared contact'
from drop-down. Presently, only searchRequest is fired in this scenario .

Fix :- When the user chooses the option 'personal and shared contact' from the dropdown, from the searchResponse, we get the
mount point of shared contact groups, and with these we fire batchRequest 'getContactRequest'. Then, we parse the 'contactResponse', update the shared
contact list with the shared contact members at runtime, and store them in cache .

https://bugzilla.zimbra.com/show_bug.cgi?id=94476

Copied from Perforce
 Change: 528526