Integrated the following from //depot/zimbra/JUDASPRIEST/...: Change 529227 by vbellows@vbellows-JP on 2014/11/18 15:51:02
Bug: 94520 -  [JA] Calendar: Suggestion Preferences pop-up window too wide to run over the screen size

Different issue - there is no maximum size specified for the Suggestions Preferences dialog, and the Japanese text is quite verbose, so the dialog expands to accomodate the text.
I've set a max-width on the dialog: on my Chome, English was at 581px, so I've set the dialog max-width to 600px.

https://bugzilla.zimbra.com/show_bug.cgi?id=94520

Copied from Perforce
 Change: 529341