Integrated the following from //depot/zimbra/JUDASPRIEST/...: Change 529185 by vbellows@vbellows-JP on 2014/11/18 10:36:05
Bug: 96356 -  Compose mail not working in IE-8 if "Contacts" tab is disabled. 

Multiple problems.  The reported one is that IE8 blows up if contacts are not enabled when trying to do a 'New Message'.
- It turns out if GAL is enabled, but not contacts, no autoComplete controls were really being setup (no handlers, causing the IE8 script error).
I've altered it to just use the normal autoComplete.  It should only be populated by GAL entries.  If contacts were enabled at one time, the user will have access to those for autocomplete.  It doesn't look like autoComplete was ever working previously in this case, despite the comments.
- if GAL is not enabled, there's no autoComplete, and the addressInput fields are properly sized now (they were being set to 30px, making for a pretty tiny typing area).

https://bugzilla.zimbra.com/show_bug.cgi?id=96356

Copied from Perforce
 Change: 529357