Bug: 96616 - Calendar search results in JS error
Some tooltips are callbacks (if they need to make a server
call to get data), so make sure the arg to stripTags() is a
string.

The tooltip is for accessibility and will not be present.
We can fix that when we move on to making Calendar
accessible.

https://bugzilla.zimbra.com/show_bug.cgi?id=96616

Copied from Perforce
 Change: 529932