Bug 96583
new message in an old conversation does not cause conversation expansion icon to appear


Some related bugs, are bug 88837, bug 87861, and bug 91083.

Long story short, my solution is in _getDisplayedMsgCount, if the msgFolder hash does not contain enough information (say it contains only a folder of ONE message but the numMsgs of the conv is 2), just use the numMsgs instead of filtering.  False positive is better than false negative in this case. The worst case scenario as I see it is that the user will get an "expand" icon, but when clicking on it, only one message will show (if the other one(s) is/are in trash). This is a rare case, I believe (considering the circumstances of this bug in addition).

This solution is not fixing the "root" cause of not having the msg folder information in some cases, but it sounds like in case of old convs or single message convs, that are not loaded in the cache, we don't have this information easily from the server, so this solution avoids having to do acrobatics to get it.

I also included an "if" wrapper around some code that seemed suspicious at first, but I don't think is related to this bug eventually. But I kept it (in ZmConvListView.prototype._changeListener).

https://bugzilla.zimbra.com/show_bug.cgi?id=96583

Copied from Perforce
 Change: 530698