Bug: 96716 - New message always added to bottom of conv, despite setting order as new-to-old
For accessibility, we interposed a container DIV to hold
each msg view. That container DIV was being appended to
the messages list, so new msg views were always added at
the bottom. The fix is to use the index to add the container
DIV to the right spot in the messages list.

Additionally, these container DIVs were not being cleaned
up, so they accumulated and threw off the indexes, causing
new messages to be added at the top even if the user is
ordering messages from old to new. The container DIVs are
now removed during reset().

Fix the msg count in the aria-label. It was skipping (1 of 15,
3 of 15, etc) since we were adding the ID to this._msgViewList
twice.

https://bugzilla.zimbra.com/show_bug.cgi?id=96716

Copied from Perforce
 Change: 530905