Bug: 96724
TinyMCE editor is created based on the textarea and it’s display and visibility property is set as none and hidden respectively for HTML mode. This is done to make the editor rendering smooth. To fix this issue visibility property is completely removed and only display property is used. 

Cause: For non-English locales, TinyMCE is loading its localization script file before rendering the editor in the tinymce.init method. During the loading time text area’s display is set as “none” which makes the TinyMCE’s internal hidden property to incorrect value.

  tinyMCE.init(tinyMCEInitObj);
  var textEl = document.getElementById(id);
  textEl.style.visibility = 'visible';
  Dwt.setVisible(textEl, this._mode === Dwt.TEXT);

https://bugzilla.zimbra.com/show_bug.cgi?id=96724

Copied from Perforce
 Change: 530941