bug 96134 - make screen readers announce status messages
(Merging using main_to_JUDASPRIEST)

We do this by making the toast itself an alert and a live region, and
by assigning an alternate text to the toast icon. In order to make
this work reliably, we clear the toast contents when hiding it;
otherwise we wouldn't announce, say, multiple draft saves.

Please note that this differs from the approach used during the
Comcast effort. Back then, we expected to announce many other messages
to the user, so we added a notion of a 'shared' live region which we
used for all notices, including toasts. This would have allowed us to
use a single 'log' for all notices -- if it weren't for the fact that
JAWS always reads entire live regions out loud. So we had different
approaches for different browsers.

Simply making the toast live has the advantage that we avoid adding
new AT-only infrastructure. Instead, the toast simply gets read out
loud.

The main other use of live regions will be autocomplete, but I think
we'd be better off by handling that as a separate issue.

(As an additional minor cleanup, move the toast and status view to
using DwtControl methods for affecting its element rather than Dwt
methods.)

Tested in Safari using VoiceOver as well as Firefox and Internet
Explorer 11 using NVDA and JAWS.

https://bugzilla.zimbra.com/show_bug.cgi?id=96134

Copied from Perforce
 Change: 531146