Bug 96733
Draft count gets increase after clicking on Cancel from Reply/Fwd compose window


This is a regression from bug 96706.

The change to ZmHtmlEditor.prototype.clear, of not clearing the textarea if editor is not defined.

Reverting bug 96706 and added a specific check for the case where we really don't want to clear the textarea.

https://bugzilla.zimbra.com/show_bug.cgi?id=96733

Copied from Perforce
 Change: 531324