Bug 96587 - Search broken, shows "Script Error: ZmItem.RESULTS_LIST[n] is not a function - Startup1_2_all.js:6423"
This seems like an intermittent issue on dogfood — I was able to verify it at first when I started working on it (and it wasn't a problem locally), but now it's gone from dogfood too.

Added a function that does the check if ZmItem.RESULTS_LIST[type] is a funciont and if not, returns a new ZmList.

Added a debug message with params info for cases when something is wrong with the type.

https://bugzilla.zimbra.com/show_bug.cgi?id=96587

Copied from Perforce
 Change: 531548