Bug 96847
collapse/expand arrow is selecting the item in folder chooser dialog/drop-down


This deals with 2 issues:
1. The "Mail Folders" root/header in the search tab advanced filter folder chooser dialog, should not be selectable. Easy fix in ZmFolderSearchFilter.prototype._getMoveParams

2. the collapse/expand arrow should only do the collapse/expand, and not do selection. For this I added a simple _expandable field to DwtTreeItem to indicate if item is expandable (if it has children OR some forceNode param, that I think is never passed, is true in DwtTreeItem.prototype._initializ)
More importantly, since DwtTreeItem._nodeIconMouseDownHdlr takes care of collapsing/expanding the node when clicking on the arrow, we do not need to do anything in DwtTreeItem._mouseUpListener if the click is ON the arrow. Thus checking for that case and exiting from that method if that's the case.

https://bugzilla.zimbra.com/show_bug.cgi?id=96847

Copied from Perforce
 Change: 531695