Bug 86122
Buttons on Address Picker are out of the window


There are 2 options to size things up in the contact picker.
1. (the previous one) - size the chooser (say to 300) and let everything else including the entire dialog size to the content size. due to the issues with child windows, resizing of the chooser is required with different settings of detailed search, etc.

2. what I changed to now - size the dialog as a whole - and calculate the size of the chooser as the dialog height minus all the constant sections stuff (the search field table, the paging buttons, the "ok"/"cancel" buttons, etc).

I chose option 2 since it provides the following advantages:
a. we control the height of the dialog directly, for parent and child windows cases, no surprises.
b. the height of the dialog stays constant, regardless of the other elements that get added/removed, such as the case when switching (in detailed search case) from GAL search to Contacts (the "Department" gets added only for GAL for some reason). It would also work in case phonetic name search is set to true, which would add another row/field.

This solution is more straight forward I think, easier to understand.

Another fix was un-remove of the searchIcon div in case of detailed search, as it provided a feedback to user when search is running, AND caused an exception when removed. I consulted with Josh and he suggested moving it to after the search button, so it can be there for both regular and detailed search, and we don't have that awkward icon in a separate row for detailed search. Josh suggested I only show that animated one while searching, and remove the magnifying glass icon otherwise, but I think I like keeping that magnifying glass icon when not searching. It's nice and emphasizes the role of those fields.

As for sizing the dialog, I chose 460 as height for the main window (since 462 happened to be what it was implicitly set to previously), and 460 - 100 for the child window. It seems to work well.

https://bugzilla.zimbra.com/show_bug.cgi?id=86122

Copied from Perforce
 Change: 532162