Bug 96973
XSS vulnerability in Find Shares dialog if one of the shared folder contains suspicious html/js code


encode the evil HTML when creating the tree item text. I wish we could encode it in the lower level DwtTreeItem.prototype.setText code, but it sets innerHTML on the text cell, and so perhaps there are callers that pass legit HTML (say bold text or something).

the small thing is ZmShareSearchDialog is just an allergy I have for null checking collections/arrays. In this case for sure it can never be null, only empty.

https://bugzilla.zimbra.com/show_bug.cgi?id=96973

Copied from Perforce
 Change: 532298