BUG 96142 Added tabbing in edit folder properties dialog.
(Merging using main_to_JUDASPRIEST)

The DwtPropertySheet -- used for name, type, color, etc. in the dialog --
now adds every DOM child to its tabgroup. In addition, labels' for attribute
refers to the id of the value inserted.

A tabgroup is added to ZmFolder(Property|Retention)View through
the DwtPropertyPage class. Concrete members of it is added in the two views.

https://bugzilla.zimbra.com/show_bug.cgi?id=96142

Copied from Perforce
 Change: 532356