Bug: 96999 - Location name and other information not appearing in the Find location dialog.
For Bug 96792, not all attributes were included in ZmContact.ATTR_VARIANTS, resulting in attributes being stripped out in ZmContact.mapAttrs.  Rather than adding this set of missing attributes (which have no variants), I've made it so that if a mapping doesn't exist, the attribute will be simply be passed along.  So if there are any other missing attributes (which I'm guessing is more likely than improper mapping as in 96792) they will be left intact 

https://bugzilla.zimbra.com/show_bug.cgi?id=96999

Copied from Perforce
 Change: 532514