Bug 94226 - Unnecessary scrollbar appeared for attached message Changeset: 1. ZmHtmlEditor.js
Cause ::-The issue here is whenever the editor loads, editor iframe body has scrollHeight which is approx. 16px greater than editor's iframe body  offsetHeight, that causes the scrollbar to appear, in this case or even when the new editor is initialised for first time.

Fix :-We are checking, if the editor has been initialised and if editor’s iframe body scrollHeight - offsetHeight  ( approx 16 px is comes by default, when editor is initialised ), we increase the body offsetHeight to avoid scrolling .

https://bugzilla.zimbra.com/show_bug.cgi?id=94226

Copied from Perforce
 Change: 532645