Bug 97067
ZWC: Slow search results


Like I mentioned in the bug, this is a problem in the case of preview pane at the bottom (multi-column) and not on the right.

Anyway, the methods that show as most time consuming on the Chrome timing, are the native DOM ones such as get clientWidth and so on. That's probably expected. But I believe the reason for the issue is not those methods, but rather that they are called too many times (Chrome shows the total CPU of all calls to each method). I was able to see a couple places where code that ends up calling DwtListView.prototype._resetColWidth (which is one of the issues in the Chrome timing, getting clientWidth from DOM), is called in a loop. _fitToContainer is called many times from a path that includes ZmSkin._reflowApp.

ZmSkin._reflowApp is called multiple times due to a couple of loops in ZmAppViewMgr where it's not really required to be called at least until the last iteration of the loop (loops over components). So my fix is to tackle this issue.

Interestingly ZmAppViewMgr.prototype.showSkinElement already has a param, that might not be used (could't find a use but I might miss something), noReflow which is exactly for that - to pass to ZmSkin.show not to reflow all the components. So I used that.

https://bugzilla.zimbra.com/show_bug.cgi?id=97067

Copied from Perforce
 Change: 532961