Integrate change 533361 from JUDASPRIEST to main
-------------
Bug 96748
Acceptance mail is not sent when user accepts the accepted new proposed time invite from organizer
Bug 38179 added a check against ZmMailListController.REPLY_ACTION_MAP, which I'm not sure was really needed, and definitely does not seem needed now.
Bug 39490 inlined that map (copy paste pretty much) to the method, due to some JS bug in new window (ZmMailListController was not loaded). But that's just a historical point, the bug existed ever since bug 38179.
The problem was that the map did not include the ZmOperation.REPLY_ACCEPT as key, even though that's what is passed in case the user clicks "accept" without choosing "notify organizer" from the button dropdown. But the "notify organizer" case is the default one so it should have acted the same.
Anyway the way I see it, this mapping check is not required. At first I thought the only case that's covered by that check (i.e. legitimately should not be in the map) was the ZmOperation.REPLY_NEW_TIME case, but that case never happens here - if user proposes a new time, it goes through the compose invite view, and does not go through this code. Interestingly that case in the switch statement at the top of ZmMailMsg.prototype._sendInviteReply, has a comment " // XXX: WRONG MAPPING!" which is true, this comment (originally said "// ?? IS THIS MAPPING RIGHT") survived since like 2005. :)  I decided to finally remove this wrong mapping.

https://bugzilla.zimbra.com/show_bug.cgi?id=96748

Copied from Perforce
 Change: 533362