Bug: 97231,97238 - Consolidate code that checks preconditions
Create a general-use function for checking preconditions where
a precondition can be a setting, a boolean, a function, or an
array (of preconditions). There was code for doing that to some
degree in five different places. Instead of creating a sixth for
ZmOperation, I refactored into ZmAppCtxt to replace the others. If a
list is passed, it defaults to requiring that all its members
be true. A flag can be passed to change that to any of them being
true.

- Clean up preconditions in various places to make them consistent
- Fix a bug in filters where the flagging action (under "Mark As")
was listed even with flagging disabled.
- Fix similar bugs where flagging showed up in search tab filters and
in search box autocomplete for "is:".
- Fix wonky formatting of ZmSearchAutocomplete
- Delete unused file ZmPrefSection.js
- Give ZmSettings a "set" function since it already has "get", and
use it instead of calling setValue() on the ZmSetting

https://bugzilla.zimbra.com/show_bug.cgi?id=97231
https://bugzilla.zimbra.com/show_bug.cgi?id=97238

Copied from Perforce
 Change: 533685