Bug 97098
Replies to conversations with drafts should not quote the draft


We recently made a change to display the drafts in non-draft folder views (i.e. Inbox). So we removed ID_DRAFTS from ZmMailApp.FOLDERS_TO_OMIT. This is good for many use-cases (such as counting how many messages are displayed, expanding the conv, etc), but for the case of replying/forwarding/replying-all, we still need to omit Drafts (for all cases). So I added a method ZmMailApp.getReplyFoldersToOmit that always adds the drafts folder to what's returned from getFoldersToOmit.

Note that there's a difference between the drafts folder (for reply/forward) and the other omitted folders (trash and spam,  though honestly spam should be removed from there since the server makes it so spam is never in a conversation, so it's irrelevant) - the other folders are sometimes not omitted (in trash folder view, and in search view), but the drafts folder should always be omitted in case of reply/reply-all/forward, even in the drafts folder view (though that case is hypothetical since the Drafts folder view has "reply" and "forward" disabled, even if the user changes to Conv view from the default Msg view...)

I removed ZmConvListController.prototype._forwardListener since it was an exact copy of what it overridden.

And as for ZmConv.prototype._handleResponseGetFirstHotMsg - this is weird, no idea how the foldersToOmit param was not passed to getFirstHit, sounds like a more general bug than just what's required for this to work (if the conv is not fully loaded, I think that's the case).

https://bugzilla.zimbra.com/show_bug.cgi?id=97098

Copied from Perforce
 Change: 533760